feat: refactored `SmartStoreXSearch` to interface and impl
